pyproject.toml
[project]
name = "tracer-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=23.2.0",
    "bcrypt>=4.0.0",
    "colorlog>=6.8.0",
    "dataclasses-json>=0.6.0,<1.0.0",
    "fastapi>=0.109.0,<0.110.0",
    "fr24sdk>=0.1.0",
    "geopy>=2.4.0",
    "google-genai>=0.2.0,<1.0.0",
    "httpx>=0.26.0,<1.0.0",
    "numpy>=1.24.0,<2.0.0",
    "openai>=1.10.0,<2.0.0",
    "orjson>=3.9.0,<4.0.0",
    "pandas>=2.1.0,<3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.2.0,<11.0.0",
    "psutil>=5.9.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "pydantic[email]>=2.5.0,<3.0.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0.0",
    "requests>=2.31.0,<3.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "scipy>=1.11.0,<2.0.0",
    "shapely>=2.1.2",
    "staticmap>=0.5.7",
    "torch>=2.1.0,<3.0.0",
    "uvicorn[standard]>=0.27.0,<0.28.0",
    "xgboost==3.1.2",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
